Fix permissions loading
